feat: Telegram (#2)
* feat(settings): add Deno support for TypeScript and TSX

Adds configuration for Deno as the language server for TypeScript and TSX. 
This change enables Deno support and sets it as the primary formatter, 
improving development experience and ensuring compatibility with Deno's 
features.

* feat: add .env file and update dependencies in deno.lock

Add a .env file to manage environment variables. Update the 
deno.lock file to include the latest versions of dependencies, 
ensuring compatibility and access to new features and fixes. 
This enhances the project's maintainability and performance.

* feat: update .gitignore to exclude Copycat files

Add entries for Copycat and Copycat.exe to the .gitignore file.  
This prevents these files from being tracked in the repository,  
ensuring a cleaner project structure and avoiding unnecessary  
clutter in version control.